requirements.txt
google-generativeai
python-dotenv
Flask
flask-cors
pydub
gunicorn
ffmpeg-python
edge-tts
psycopg2-binary